Print memory use for all resources in ResourceCache::PrintMemoryUsage(). Change memory budget & use API to use unsigned long long (64bit). Expose GetFileSizeString() to script. Refactor Engine to use PrintMemoryUsage(). Match width of profiler and resource outputs. Rename Profiler::GetData() to Profiler::PrintData() to match the new ResourceCache function.

+++ Source/Urho3D/IO/FileSystem.cpp
@@ -1069,7 +1069,7 @@ bool IsAbsolutePath(const String& pathName)
 }
 
 /// Convert a memory size into a formatted size string, of the style "1.5 Mb"
-String GetFileSizeString(unsigned long memorySize)
+String GetFileSizeString(unsigned long long memorySize)
 {
     const char* memorySizeStrings = "kMGTPE";
